Start pulling test code into separate modules
