+++ .eslintrc.js
@@ -5,7 +5,9 @@ module.exports = {
     },
     "plugins": ["@typescript-eslint"],
     "parserOptions": {
-        "project": "tsconfig.json",
         "ecmaVersion": 2018
+    },
+    rules: {
+        'no-console': 'off'
     }
 };
