+++ package.json
@@ -4,17 +4,17 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "compile": "tsc && for file in $(find dist -name '*.js' -o -name '*.map'); do NEWFILENAME=`echo \"$file\" | sed 's/.js$/.mjs/g; s/.js.map$/.mjs.map/g'`; mv \"$file\" $NEWFILENAME; sed -i 's/.js/.mjs/g' $NEWFILENAME; done",
+    "compile": "tsc && for file in $(find dist -name '*.js' -o -name '*.map'); do NEWFILENAME=`echo \"$file\" | sed 's/.js$/.mjs/g; s/.js.map$/.mjs.map/g'`; mv \"$file\" $NEWFILENAME; done",
     "compile-watch": "tsc-watch --onsuccess \"node dist/index.js\"",
-    "lint": "eslint -c .eslintrc.js  src",
+    "lint": "eslint -c .eslintrc.js src",
     "test:e2e": "echo \"Error: no e2e test specified, i.e. run with cmd node e2e \" && exit 1",
     "test:unit": "jest",
     "test:unit-watch": "jest --watchAll",
     "test": "npm run test:unit && npm run test:e2e",
     "deploy": "clasp push",
     "build": "npm run lint && npm run compile",
-    "start": "npm run build && node dist/index.mjs",
-    "debug": "npm run compile && node --inspect dist/index.mjs"
+    "start": "npm run build && node --experimental-specifier-resolution=node dist/index.mjs",
+    "debug": "npm run compile && node --experimental-specifier-resolution=node --inspect dist/index.mjs"
   },
   "keywords": [],
   "author": "",
